agents/seo/unified-llm-scraper/requirements.txt
requests==2.32.5
python-dotenv==1.2.1
streamlit==1.53.0
supabase==2.27.2